#43 Integration Tests will be place into 'integration' package
